+++ src/player/brain/brainstates/brainstate_cover.cpp
@@ -9,7 +9,7 @@ namespace SenselessSoccer {
 // ------------------------------------------------------------
 // BrainCover
 // ------------------------------------------------------------
-BrainCover::BrainCover(Player *p) : BrainState(p) {
+BrainCover::BrainCover (Player *p) : BrainState (p) {
     last_ball_sector = last_target_sector = -1;
 }
 
@@ -23,53 +23,53 @@ void BrainCover::OnStart() {
 // ------------------------------------------------------------
 // OnStep
 // ------------------------------------------------------------
-void BrainCover::OnStep(const double dt) {
+void BrainCover::OnStep (const float dt) {
     GameLib::Vector3 destination;
 
     // get the current ball sector
-    int ball_sector = player->pitch->grid.PointToSector(player->ball->GetPosition());
+    int ball_sector = player->pitch->grid.PointToSector (player->ball->GetPosition());
 
     // if ball sector is valid and has changed
-    if(ball_sector >= 0 && ball_sector != last_ball_sector) {
+    if (ball_sector >= 0 && ball_sector != last_ball_sector) {
 
         // get the new target sector depending on player role
-        int target_sector = player->role->GetPosition(ball_sector, player->my_team->side);
+        int target_sector = player->role->GetPosition (ball_sector, player->my_team->side);
 
         // modify based on tactical instructions?
         int m = 0;
 
-        if(modifiers.size()) {
-            switch(modifiers[0]) {
+        if (modifiers.size()) {
+            switch (modifiers[0]) {
 
-            case COVER_PRESS:
-                m = -1;
-                break;
+                case COVER_PRESS:
+                    m = -1;
+                    break;
 
-            case COVER_PRESS_HIGH:
-                m = -2;
-                break;
+                case COVER_PRESS_HIGH:
+                    m = -2;
+                    break;
 
-            case COVER_SHADOW:
-                m = 1;
-                break;
+                case COVER_SHADOW:
+                    m = 1;
+                    break;
 
-            case COVER_SHADOW_DEEP:
-                m = 2;
-                break;
+                case COVER_SHADOW_DEEP:
+                    m = 2;
+                    break;
 
-            case MODIFIER_NONE:
-            default:
-                break;
+                case MODIFIER_NONE:
+                default:
+                    break;
             }
         }
 
-        target_sector = player->pitch->grid.OffsetSectorY(target_sector, m);
+        target_sector = player->pitch->grid.OffsetSectorY (target_sector, m);
 
         // only if the target sector has changed
-        if(target_sector != last_target_sector) {
+        if (target_sector != last_target_sector) {
             last_target_sector = target_sector;
-            destination = player->pitch->grid.GetSectorCenter(target_sector).vector();
-            player->brain.locomotion.ActivateArrive(destination);
+            destination = player->pitch->grid.GetSectorCenter (target_sector).vector();
+            player->brain.locomotion.ActivateArrive (destination);
         }
 
         // save for next time round
@@ -79,7 +79,7 @@ void BrainCover::OnStep(const double dt) {
     // check stop condition
     GameLib::Vector3 distance = player->physical->position - destination;
 
-    if(distance.magnitude() < 10) {
+    if (distance.magnitude() < 10) {
         player->brain.locomotion.Cancel();
     }
 }
@@ -100,9 +100,9 @@ bool BrainCover::StateOver() {
 // ------------------------------------------------------------
 // Modify
 // ------------------------------------------------------------
-void BrainCover::Modify(modifier mod) {
+void BrainCover::Modify (modifier mod) {
     modifiers.clear();
-    modifiers.push_back(mod);
+    modifiers.push_back (mod);
     last_ball_sector = -1;
 }
