+++ src/player/brain/brainstates/brainstate_cover.h
@@ -1,8 +1,8 @@
 /**
- * \file 	src/player/locomotion/BrainCover.h
- * \author 	Paul Higgins
- * \date 	14.08.2017
- * @brief 	BrainCover type locomotion
+ * \file    src/player/locomotion/BrainCover.h
+ * \author  Paul Higgins
+ * \date    14.08.2017
+ * @brief   BrainCover type locomotion
  *
  * Description
  *
@@ -28,7 +28,7 @@ public:
      * @brief constructor
      * @param [in] p pointer back to player for state machine context
      */
-    explicit BrainCover(Player *p);
+    explicit BrainCover (Player *p);
 
     /**
      * @brief state on start
@@ -39,7 +39,7 @@ public:
      * @brief state on step
      * @param [in] dt time delta
      */
-    virtual void OnStep(const double dt) override;
+    virtual void OnStep (const float dt) override;
 
     /**
      * @brief state on end
@@ -54,7 +54,7 @@ public:
     /**
      * @brief modify the state parameters on the fly
      */
-    virtual void Modify(modifier mod) override;
+    virtual void Modify (modifier mod) override;
 
 protected:
     /// track last ball sector
