+++ src/player/brain/brainstates/brainstate_dribble.cpp
@@ -9,7 +9,7 @@ namespace SenselessSoccer {
 // ------------------------------------------------------------
 // Constructor
 // ------------------------------------------------------------
-BrainDribble::BrainDribble(Player *p) : BrainState(p) {
+BrainDribble::BrainDribble (Player *p) : BrainState (p) {
 }
 
 // ------------------------------------------------------------
@@ -19,7 +19,7 @@ void BrainDribble::OnStart() {
     player->brain.statename = "DRIBBLE";
     player->OnGainedPossession();
     player->velocity = player->last_direction;
-    assert(player->velocity.magnitude());
+    assert (player->velocity.magnitude());
     change_direction();
     timer.Start();
 }
@@ -27,16 +27,16 @@ void BrainDribble::OnStart() {
 // ------------------------------------------------------------
 // OnStep
 // ------------------------------------------------------------
-void BrainDribble::OnStep(const double _dt) {
+void BrainDribble::OnStep (const float _dt) {
 
-    if(!player->brain.in_pitch(_dt)) {
+    if (!player->brain.in_pitch (_dt)) {
         /*********************************************
         * stay in pitch
         * *******************************************/
-        GameLib::Vector3 new_direction = player->velocity.rotated(45).roundAngle(45);
-        player->brain.locomotion.ActivateHead(new_direction);
+        GameLib::Vector3 new_direction = player->velocity.rotated (45).roundAngle (45);
+        player->brain.locomotion.ActivateHead (new_direction);
 
-    } else if(timer.GetTicks() > 2000) {
+    } else if (timer.GetTicks() > 2000) {
         /*********************************************
         * test random change dir
         * *******************************************/
@@ -57,22 +57,22 @@ void BrainDribble::OnEnd() {
 // StateOver
 // ------------------------------------------------------------
 bool BrainDribble::StateOver() {
-    if(!player->ball_under_control()) {
+    if (!player->ball_under_control()) {
         next_state = BRAIN_SUPPORT;
         return true;
     }
 
-    else if(!pass_timer_started && player->my_team->key_players.short_pass_candidates.size()){
+    else if (!pass_timer_started && player->my_team->key_players.short_pass_candidates.size()) {
 //         pass_timer_started = true;
 //         pass_timer.Start();
 //         next_state = BRAIN_PASS;
     }
 
-    else if(pass_timer_started && pass_timer.GetTicks()>100){
+    else if (pass_timer_started && pass_timer.GetTicks() > 100) {
         return true;
     }
 
-    else if(player->distance_to_goal() < 650){
+    else if (player->distance_to_goal() < 650) {
         next_state = BRAIN_SHOOT;
         return true;
     }
@@ -84,28 +84,28 @@ bool BrainDribble::StateOver() {
 // change_direction
 // ------------------------------------------------------------
 void BrainDribble::change_direction() {
-    GameLib::Vector3 new_direction = player->velocity.rotated(rand() % 2 == 1 ? 45 : -45).roundAngle(45);
+    GameLib::Vector3 new_direction = player->velocity.rotated (rand() % 2 == 1 ? 45 : -45).roundAngle (45);
     player->brain.locomotion.Cancel();
-    player->brain.locomotion.ActivateHead(new_direction);
+    player->brain.locomotion.ActivateHead (new_direction);
 }
 
 // ------------------------------------------------------------
 // Modify
 // ------------------------------------------------------------
-void BrainDribble::Modify(modifier mod) {
+void BrainDribble::Modify (modifier mod) {
 
-    switch(mod) {
-    case DRIBBLE_CHANGE_DIRECTION:
-        change_direction();
-        break;
+    switch (mod) {
+        case DRIBBLE_CHANGE_DIRECTION:
+            change_direction();
+            break;
     }
 }
 
 // ------------------------------------------------------------
 // Init
 // ------------------------------------------------------------
-void BrainDribble::Init(Compass dir) {
+void BrainDribble::Init (Compass dir) {
     timer.Start();
-    player->brain.locomotion.ActivateHead(Metrics::compasstoVector(dir));
+    player->brain.locomotion.ActivateHead (Metrics::compasstoVector (dir));
 }
 }
