+++ src/player/brain/brainstates/brainstate_getball.cpp
@@ -5,7 +5,7 @@ namespace SenselessSoccer {
 // ------------------------------------------------------------
 // BrainGetBall
 // ------------------------------------------------------------
-BrainGetBall::BrainGetBall(Player *player) : BrainState(player) {
+BrainGetBall::BrainGetBall (Player *player) : BrainState (player) {
 }
 
 // ------------------------------------------------------------
@@ -13,13 +13,13 @@ BrainGetBall::BrainGetBall(Player *player) : BrainState(player) {
 // ------------------------------------------------------------
 void BrainGetBall::OnStart() {
     player->brain.statename = "GET BALL";
-    player->brain.locomotion.ActivatePursue(player->ball->physical);
+    player->brain.locomotion.ActivatePursue (player->ball->physical);
 }
 
 // ------------------------------------------------------------
 // OnStep
 // ------------------------------------------------------------
-void BrainGetBall::OnStep(const double _dt) {
+void BrainGetBall::OnStep (const float _dt) {
 }
 
 // ------------------------------------------------------------
@@ -33,7 +33,7 @@ void BrainGetBall::OnEnd() {
 // StateOver
 // ------------------------------------------------------------
 bool BrainGetBall::StateOver() {
-    if(player->ball_under_control()){
+    if (player->ball_under_control()) {
         next_state = BRAIN_DRIBBLE;
         return true;
     }
