+++ src/player/brain/brainstates/brainstate_idle.h
@@ -11,7 +11,7 @@ public:
      * @brief BrainIdle
      * @param player
      */
-    explicit BrainIdle(Player *player);
+    explicit BrainIdle (Player *player);
 
     /**
      * @brief OnStart
@@ -22,7 +22,7 @@ public:
      * @brief OnStep
      * @param _dt
      */
-    virtual void OnStep(const double _dt) override;
+    virtual void OnStep (const float _dt) override;
 
     /**
      * @brief OnEnd
