+++ src/player/brain/brainstates/brainstate_pass.h
@@ -10,7 +10,7 @@ public:
      * @brief BrainPass
      * @param p player
      */
-    explicit BrainPass(Player *p);
+    explicit BrainPass (Player *p);
 
     /**
      * @brief OnStart
@@ -21,7 +21,7 @@ public:
      * @brief OnStep
      * @param _dt time delta
      */
-    virtual void OnStep(const double _dt) override;
+    virtual void OnStep (const float _dt) override;
 
     /**
      * @brief OnEnd
@@ -37,12 +37,12 @@ public:
      * @brief init the pass state
      * @param receiver receiving player
      */
-    void Init(Player *receiver);
+    void Init (Player *receiver);
 
     /**
     * @brief modify the state parameters on the fly
     */
-    virtual void Modify(modifier mod) override;
+    virtual void Modify (modifier mod) override;
 
 protected:
