+++ src/player/brain/brainstates/brainstate_press.cpp
@@ -5,7 +5,7 @@ namespace SenselessSoccer {
 // ------------------------------------------------------------
 // BrainPress
 // ------------------------------------------------------------
-BrainPress::BrainPress(Player *player) : BrainState(player) {
+BrainPress::BrainPress (Player *player) : BrainState (player) {
 }
 
 // ------------------------------------------------------------
@@ -13,13 +13,13 @@ BrainPress::BrainPress(Player *player) : BrainState(player) {
 // ------------------------------------------------------------
 void BrainPress::OnStart() {
     player->brain.statename = "PRESS";
-    player->brain.locomotion.ActivateIntercept(player->ball->physical);
+    player->brain.locomotion.ActivateIntercept (player->ball->physical);
 }
 
 // ------------------------------------------------------------
 // OnStep
 // ------------------------------------------------------------
-void BrainPress::OnStep(const double _dt) {
+void BrainPress::OnStep (const float _dt) {
 }
 
 // ------------------------------------------------------------
@@ -36,7 +36,7 @@ bool BrainPress::StateOver() {
 
     // TODO: end condition: got ball, or decision to stop pressing, or decision to slide etc
 
-    if(player->ball_under_control()){
+    if (player->ball_under_control()) {
         next_state = BRAIN_DRIBBLE;
         return true;
     }
